pyproject.toml
[build-system]
    build-backend = "setuptools.build_meta"
    requires      = ["setuptools>=77.0.3"]

[project]
    authors = [
        { name = "Blackteahamburger", email = "blackteahamburger@outlook.com" },
    ]
    classifiers = [
        "Development Status :: 5 - Production/Stable",
        "Environment :: Console",
        "Intended Audience :: Developers",
        "Intended Audience :: Education",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: POSIX",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Topic :: Education",
        "Topic :: Software Development :: Embedded Systems",
    ]
    dependencies = ["pyserial>=3.0.1,<4.0"]
    description = "A module and utility to work with the simple filesystem on the BBC micro:bit"
    license = "MIT"
    license-files = ["AUTHORS", "LICENSE"]
    maintainers = [
        { name = "Blackteahamburger", email = "blackteahamburger@outlook.com" },
    ]
    name = "microfs2"
    readme = "README.rst"
    requires-python = ">=3.11"
    version = "2.0.6"

    [project.optional-dependencies]
        dev = ["coverage", "pytest-cov", "pytest>=5", "ruff", "sphinx"]

    [project.urls]
        Homepage = "https://github.com/blackteahamburger/microfs"

    [project.scripts]
        ufs = "microfs.main:main"

[tool.coverage.run]
    omit = ["__main__.py"]
